Moved resources related things to their own folder
