+++ src/components/Header.tsx
@@ -1,6 +1,6 @@
-import { Logo } from "./Logo";
+import Logo from "./Logo";
 
-export default function Header() {
+export function Header() {
 	return <header
 		className="sticky top-0 z-40 flex-none mx-auto w-full bg-white md:bg-white/90 dark:bg-slate-900 dark:md:bg-slate-900/90 md:backdrop-blur-sm border-b dark:border-b-0"
 	>
@@ -153,3 +153,5 @@ export default function Header() {
 		</div>
 	</header>;
 }
+
+export default Header;
