+++ src/components/Home.tsx
@@ -2,7 +2,7 @@ interface Props {
 	loginURL: string
 }
 
-export default function Home({ loginURL }: Props) {
+export function Home({ loginURL }: Props) {
 	return <section>
 		<div className="max-w-6xl mx-auto px-4 sm:px-6">
 			<div className="py-12 md:py-20">
@@ -29,3 +29,5 @@ export default function Home({ loginURL }: Props) {
 		</div>
 	</section>
 }
+
+export default Home;
