+++ src/components/resources/Resource.tsx
