+++ src/models/Resource.ts
@@ -1,7 +1,9 @@
-export default interface Resource {
+export interface Resource {
 	id: number;
 	author: string;
 	title: string;
 	description:string;
 	filename: string;
 }
+
+export default Resource;
