+++ src/pages/Resources.astro
@@ -1,7 +1,7 @@
 ---
 import ListOfResource from "@layouts/ListOfResource.astro";
 import Wrapper from "@layouts/Wrapper.astro";
-import { Res } from "@components/data";
+import { Res } from "@components/resources/data";
 ---
 
 <Wrapper title="Proyecto Plataforma v2">
